[ET][Testing] Build test_backend_compiler_lib when testing is on (#9953)
Adding test_backend_compiler_lib to pybinds when tests are on. This
allows us to run some more backend tests in the OSS build. There were a
couple of other test files that weren't able to be enabled:
```
    --ignore=exir/backend/test/test_backends.py
    --ignore=exir/backend/test/test_backends_lifted.py
    --ignore=exir/backend/test/test_partitioner.py
```

This was because they were doing some
torch.ops.load_library("buck_path") or there was another backend
ExecutorBackend that was required. I'll limit the changes of this PR to
just be enablement of the BackendWithCompilerLibDemo backend.