Arm backend:  Add TOSA VGF encapsulated compilation target. (#10476)
Add TOSA VGF encapsulated compilation target.

This change enables support for "vgf" files which wrap TOSA output and
include memory planning for target devices which can JIT TOSA to the
target ISA on-device.

 - Add a VgfQuantizer (same as TOSAQuantizer)
 - Add a VgfBackend and VgfPartitioner to produce TOSA wrapped in a VGF
 - Requires yet to be released converter_backend

### Test plan
As this is a new encapsulation with a tool that's not yet released,
integration with unit tests will come in a subsequent commit.

Signed-off-by: Rob Elliott <robert.elliott@arm.com>