Fix libflatccrt race (#10918)
### Summary

Seems like there is a race in building `libflatccrt.a`. This issue has
existed for a while: https://github.com/pytorch/executorch/issues/7300.
It was temporarily mitigated in
https://github.com/pytorch/executorch/pull/7570 by just reducing the
parallelism.

In this diff I attempt to fix it. This is just my assumption of what is
wrong. Given flatccrt builds a debug version with a `_d` suffix, if the
target isn't depended on (i.e. some target don't use the conditional
target name) then the order of how the lib is built causes a race. So
for now, always use the non-debug version.

Given it's a race, I was never able to repro the issue locally — I can't
guarantee this is the problem. However, it seems my recent changes in
https://github.com/pytorch/executorch/pull/10855 has increased the
frequency of the problem in CI.

### Test plan

CI

cc @larryliu0820