Use "pybind" preset when building a wheel (#10951)
### Summary

Now that we have presets for pybind, lets use that and eliminate the
need to set cmake configurations in `setup.py`. Moreover, we now stop
relying on defaults and fully rely on the cmake configurations. This
includes relying on `CMakeCache.txt` to determine what features are
enabled to determine what extensions to copy.

### Test plan

CI +

```
$ rm -rf pip-out && python setup.py bdist_wheel
$ rm -rf pip-out && ./install_executorch.sh
```

cc @larryliu0820