Add portable rand kernel implementation (#11127)
### Summary
Add a portable operator implementation for ATen rand. This is a core op
that we previously have not had a kernel for.

### Test plan
I've added operator-level tests for rand. It's relatively sparse, but
validates that the mean and stdev are reasonably sane, as well as that
the operator functions correctly for different dtypes, ranks, and sizes.