Introduce PAL function table (#10675)
Summary:
This is part 1 of the implementation of the PAL changes described in
https://github.com/pytorch/executorch/discussions/10432. This PR
introduces a struct (`pal_table`) to hold function pointers to the PAL
function implementations. There is a singleton instance of this struct,
which is initialized with the weak/strong et_pal_ functions -
maintaining backwards compatibility with the existing override
mechanism.

I've then added wrapper functions for the PAL into the
executorch::runtime namespace which dispatch through the function table.
It is intended that callers use these functions instead of the "raw"
et_pal_ methods in order to correctly dispatch through the function
table. In the following PR, I update ET callers to do this.

Differential Revision: D74121895