Enforce required cmake flags as explicit requirements (#11457)
### Summary

The build file currently turns on a bunch of implicitly dependent flags,
perhaps unnecessarily. Let's make these dependencies explicit now.

For example, previously if `EXECUTORCH_BUILD_EXTENSION_TRAINING` was
turned on, it would then turn on its own set of dependencies. This is
good from a developer experience point of view because users would need
to just turn on one flag. However, from a build point of view, it makes
it harder to maintain and understand what flags are conflicting with one
another. Now, if you want to build the training extension, you will have
to explicitly turn on the required flags. However, with our build
presets — most developers should not have to do to this!

I have created a follow up ticket to actually comb through the existing
flags and trim the fat if necessary:
https://github.com/pytorch/executorch/issues/11458

### Test plan

CI

cc @larryliu0820