Arm backend: Add embedding.default + index_select (#11636)
aten.embedding.default is requiring int64 indices which is not supported
by TOSA. A pass is introduced to insert a int64->int32 cast. This
enables aten.embedding to be decomposed into index_select, which can
handle int32 indices and maps easier to TOSA. So support for
aten.index_select is added as well, and for now it only supports [1, x,
y] and [x, y] dimension inputs since that covers aten.embedding.default
use case.

In summary this patch adds:
- One pass for inserting a cast, which is done very early since it needs
to be rejected until TOSA supports int64
- One pass for decomposing aten.embedding
- Mapping of index select to TOSA
- Unit tests for above